refactor(suite): rationalize useAccounts, use instead of adhoc logic
